+++ keyboards/rate/pistachio_pro/config.h
@@ -21,39 +21,17 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #define MATRIX_ROWS 6
 #define MATRIX_COLS 18
 
-#define MATRIX_ROW_PINS { D4, C6, D7, B3, B4, B5 }
-#define MATRIX_COL_PINS { E6, F0, F1, F4, F5, F6, F7, B6, D6 }
-
 /* Mechanical locking support. Use KC_LCAP, KC_LNUM or KC_LSCR instead in keymap */
 #define LOCKING_SUPPORT_ENABLE
 /* Locking resynchronize hack */
 #define LOCKING_RESYNC_ENABLE
 
-/*
- * Feature disable options
- *  These options are also useful to firmware size reduction.
- */
-
-/* disable debug print */
-//#define NO_DEBUG
-
-/* disable print */
-//#define NO_PRINT
-
-/* disable action features */
-//#define NO_ACTION_LAYER
-//#define NO_ACTION_TAPPING
-//#define NO_ACTION_ONESHOT
-
 /* OLED */
 #define OLED_DISPLAY_128X64
 #define OLED_FONT_H "keyboards/rate/pistachio_pro/lib/glcdfont.c"
 
-#define OLED_FONT_WIDTH 6
 #define OLED_FONT_HEIGHT 6
 
 #define OLED_UPDATE_INTERVAL 1000
 
 #define OLED_TIMEOUT 0
-
-#define OLED_BRIGHTNESS 255
