+++ keyboards/rate/pistachio_pro/info.json
@@ -3,10 +3,20 @@
     "manufacturer": "rate",
     "url": "https://rates.booth.pm/items/2925303",
     "maintainer": "rate",
+    "features": {
+      "mousekey": true,
+      "extrakey": true,
+      "oled": true,
+      "encoder": true
+    },
     "usb": {
         "vid": "0x5255",
         "pid": "0xF40C",
         "device_version": "0.0.1"
+    },
+    "matrix_pins": {
+        "cols": ["E6", "F0", "F1", "F4", "F5", "F6", "F7", "B6", "D6"],
+        "rows": ["D4", "C6", "D7", "B3", "B4", "B5"]
     },
     "encoder": {
         "rotary": [
