--- keyboards/rate/pistachio_pro/keymaps/default/readme.md
@@ -1,1 +1,0 @@
-# The default keymap for pistachio_pro
