--- keyboards/rate/pistachio_pro/keymaps/rate/readme.md
@@ -1,1 +1,0 @@
-# The rate's keymap for pistachio_pro
