--- keyboards/rate/pistachio_pro/keymaps/via/readme.md
@@ -1,1 +1,0 @@
-# The via keymap for pistachio_pro
